Update correct answer button
This commit updates the dictionary in Question.cs to match the new
implementation of the correct answer button alternating, so that the
text that is displayed on each correct or incorrect answer button
matches the correct and incorrect answer text for the game to be played
correctly. It also corrects some typos in the other script files.

Signed-off-by: Isaiah Andrade <isaiah_andrade@student.uml.edu>